refactor: View설계 단순화...
