--- src/main/java/view/inputview/InputView.java
@@ -1,9 +1,0 @@
-package view.inputview;
-
-import java.util.Set;
-
-public interface InputView {
-    int inputPurchaseMoney();
-    int inputBonusNumber();
-    Set<Integer> inputWinningNumbers();
-}
