--- src/main/java/view/outputview/OutputView.java
@@ -1,11 +1,0 @@
-package view.outputview;
-
-import java.util.List;
-import model.LottoTicketDto;
-import model.WinningResultDto;
-
-public interface OutputView {
-    void showLottoTicket(List<LottoTicketDto> lottoTickets);
-    void showWinningResult(List<WinningResultDto> winningResults);
-    void showRateOfReturn(double rateOfReturn);
-}
